Add user-facing error messages to fluvio-cluster
WIP: User errors

Clean up cluster check errors

Fix formatting for error printout

Add user error for minikube tunnel down

Apply cargo clippy and cargo fmt

Fix helm error in CLI

Remove unused modules. Apply cargo fmt

Apply cargo clippy

Apply cargo fmt

Rename extensions mod to cli

Rename extension_cmd feature flag to cli

Fix imports

Remove opt module in cluster/cli/install

Organize run command

Move root_cli into lib.rs

Create independent binary for fluvio-cluster

Minimize output from color_eyre in fluvio-cluster

Clean up cluster cli error

Move user errors closer to original error

Remove unused errors from CliError

Bump versions for fluvio-cli and fluvio-cluster

Apply cargo clippy and cargo fmt
