+++ Cargo.toml
@@ -24,7 +24,6 @@ members = [
     "tests/runner",
 ]
 
-
 # profile to make image sizer smaller
 # comment out for now
 #[profile.release]
