+++ src/cli/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "fluvio-cli"
-version = "0.4.0"
+version = "0.4.1"
 edition = "2018"
 authors = ["Fluvio Contributors <team@fluvio.io>"]
 description = "Fluvio CLI"
@@ -34,7 +34,7 @@ async-h1 = "2.1.2"
 http-types = "2.4.0"
 thiserror = "1.0.20"
 eyre = "0.6.1"
-color-eyre = "0.5.5"
+color-eyre = { version = "0.5.5", default-features = false }
 which = "4.0.2"
 sha2 = "0.9.1"
 hex = "0.4.2"
@@ -47,7 +47,7 @@ k8-config = { version = "1.3.0" }
 k8-client = { version = "3.0.0", default-features = false }
 fluvio-future = { version = "0.1.8", features = ["fs", "io", "subscriber"] }
 fluvio = { version = "0.2.3", path = "../client", default-features = false }
-fluvio-cluster = { version = "0.3.1", path = "../cluster", default-features = false, features = ["cmd_extension"] }
+fluvio-cluster = { version = "0.3.1", path = "../cluster", default-features = false, features = ["cli"] }
 fluvio-package-index = { version = "0.2.0", path = "../package-index" }
 fluvio-extension-consumer = { version = "0.1.0", path = "../extension-consumer" }
 fluvio-extension-common = { version = "0.1.0", path = "../extension-common", features = ["target"]}
