+++ src/cli/src/error.rs
@@ -1,11 +1,9 @@
 use std::io::Error as IoError;
 
 use fluvio::FluvioError;
-use fluvio_cluster::{ClusterError, CheckError};
-use fluvio_cluster::extension::ClusterCmdError;
+use fluvio_extension_common::output::OutputError;
 use fluvio_extension_consumer::ConsumerError;
-
-use crate::common::output::OutputError;
+use fluvio_cluster::cli::ClusterCliError;
 use crate::common::target::TargetError;
 
 pub type Result<T> = std::result::Result<T, CliError>;
@@ -13,72 +11,29 @@ pub type Result<T> = std::result::Result<T, CliError>;
 #[derive(thiserror::Error, Debug)]
 pub enum CliError {
     #[error(transparent)]
-    IoError {
-        #[from]
-        source: IoError,
-    },
+    IoError(#[from] IoError),
     #[error(transparent)]
-    OutputError {
-        #[from]
-        source: OutputError,
-    },
-    #[error("Consumer Error")]
-    ConsumerError {
-        #[from]
-        source: ConsumerError,
-    },
-    #[error("ClusterCommand Error")]
-    ClusterCmdError {
-        #[from]
-        source: ClusterCmdError,
-    },
-    #[error("ClusterCommand Error")]
-    TargetError {
-        #[from]
-        source: TargetError,
-    },
-    #[error("Fluvio client error")]
-    ClientError {
-        #[from]
-        source: FluvioError,
-    },
+    OutputError(#[from] OutputError),
     #[error("Fluvio cluster error")]
-    ClusterError {
-        #[from]
-        source: ClusterError,
-    },
-    #[error("Fluvio cluster pre install check error")]
-    CheckError {
-        #[from]
-        source: CheckError,
-    },
+    ClusterCliError(#[from] ClusterCliError),
+    #[error("Target Error")]
+    TargetError(#[from] TargetError),
+    #[error("Consumer Error")]
+    ConsumerError(#[from] ConsumerError),
+    #[error("Fluvio client error")]
+    ClientError(#[from] FluvioError),
     #[error("Kubernetes config error")]
-    K8ConfigError {
-        #[from]
-        source: k8_config::ConfigError,
-    },
+    K8ConfigError(#[from] k8_config::ConfigError),
     #[error("Kubernetes client error")]
-    K8ClientError {
-        #[from]
-        source: k8_client::ClientError,
-    },
+    K8ClientError(#[from] k8_client::ClientError),
+    #[error("Package index error")]
+    IndexError(#[from] fluvio_index::Error),
+    #[error("Error finding executable")]
+    WhichError(#[from] which::Error),
+    #[error(transparent)]
+    HttpError(#[from] HttpError),
     #[error("Invalid argument: {0}")]
     InvalidArg(String),
-    #[error("Package index error")]
-    IndexError {
-        #[from]
-        source: fluvio_index::Error,
-    },
-    #[error("Error finding executable")]
-    WhichError {
-        #[from]
-        source: which::Error,
-    },
-    #[error(transparent)]
-    HttpError {
-        #[from]
-        source: HttpError,
-    },
     #[error("Unknown error: {0}")]
     Other(String),
 }
@@ -90,15 +45,22 @@ pub struct HttpError {
 }
 
 impl From<http_types::Error> for CliError {
-    fn from(e: http_types::Error) -> Self {
-        CliError::HttpError {
-            source: HttpError { inner: e },
-        }
+    fn from(inner: http_types::Error) -> Self {
+        Self::HttpError(HttpError { inner })
     }
 }
 
 impl CliError {
     pub fn invalid_arg<M: Into<String>>(reason: M) -> Self {
         Self::InvalidArg(reason.into())
+    }
+
+    pub fn into_report(self) -> color_eyre::Report {
+        use color_eyre::Report;
+
+        match self {
+            CliError::ClusterCliError(cluster) => cluster.into_report(),
+            _ => Report::from(self),
+        }
     }
 }
