+++ src/cluster/Cargo.toml
@@ -1,28 +1,32 @@
 [package]
 name = "fluvio-cluster"
-version = "0.3.1"
+version = "0.3.2"
 edition = "2018"
 license = "Apache-2.0"
 authors = ["Fluvio Contributors <team@fluvio.io>"]
 repository = "https://github.com/infinyon/fluvio"
 description = "Tools for installing and managing Fluvio clusters"
 
-
 [lib]
 name = "fluvio_cluster"
 path = "src/lib.rs"
+
+[[bin]]
+name = "fluvio-cluster"
+path = "src/main.rs"
+required-features = ["cli"]
 
 [features]
 default = ["native_tls"]
 rust_tls = ["fluvio/rust_tls", "k8-client/rust_tls"]
 native_tls = ["fluvio/native_tls", "k8-client/native_tls"]
-cmd_extension = [
+cli = [
     "structopt",
+    "color-eyre",
     "fluvio-extension-common/target",
     "prettytable-rs",
     "fluvio-runner-local",
 ]
-
 
 [dependencies]
 colored = "2.0.0"
@@ -34,21 +38,22 @@ serde = "1.0.115"
 serde_json = "1.0.57"
 thiserror = "1.0.20"
 async-trait = "0.1.21"
+color-eyre = { version = "0.5.8", default-features = false, optional = true }
 structopt = { version = "0.3.16", default-features = false, optional = true }
 prettytable-rs = { version = "0.8.0", optional = true }
 
 # Fluvio dependencies
-fluvio-helm = { version = "0.1.0", git = "https://github.com/infinyon/fluvio-helm", branch = "main" }
 fluvio = { version = "0.2.3", path = "../client", default-features = false }
-fluvio-controlplane-metadata = { version = "0.2.0", path = "../controlplane-metadata", features = ["k8"] }
+fluvio-helm = { git = "https://github.com/infinyon/fluvio-helm", rev = "6e15fd2c789c33ba8bae51a369a0983e177c2f78" }
 fluvio-future = { version = "0.1.0" }
+fluvio-runner-local = { version = "0.1.0", path = "../extension-runner-local", optional = true }
+fluvio-extension-common = { version = "0.1.0", path = "../extension-common", optional = true }
+fluvio-controlplane-metadata = { version = "0.2.0", path = "../controlplane-metadata", features = ["k8"] }
 flv-util = "0.5.2"
 k8-config = { version = "1.3.0" }
 k8-client = { version = "3.0.0", default-features = false }
 k8-obj-core = "1.1.0"
 k8-obj-metadata = "1.0.0"
-fluvio-extension-common = { version = "0.1.0", path = "../extension-common", optional = true }
-fluvio-runner-local = { version = "0.1.0", path = "../extension-runner-local", optional = true }
 
 [dev-dependencies]
 fluvio-future = { version = "0.1.0", features = ["task"] }
