+++ src/cluster/src/cli/group/delete.rs
@@ -8,7 +8,7 @@ use structopt::StructOpt;
 use fluvio::Fluvio;
 use fluvio::metadata::spg::SpuGroupSpec;
 
-use crate::extension::Result;
+use crate::cli::ClusterCliError;
 
 // -----------------------------------
 // CLI Options
@@ -22,7 +22,7 @@ pub struct DeleteManagedSpuGroupOpt {
 }
 
 impl DeleteManagedSpuGroupOpt {
-    pub async fn process(self, fluvio: &Fluvio) -> Result<()> {
+    pub async fn process(self, fluvio: &Fluvio) -> Result<(), ClusterCliError> {
         let mut admin = fluvio.admin().await;
         admin.delete::<SpuGroupSpec, _>(&self.name).await?;
         Ok(())
