+++ src/cluster/src/install.rs
@@ -22,12 +22,13 @@ use k8_client::metadata::MetadataClient;
 use k8_obj_core::service::ServiceSpec;
 use k8_obj_metadata::InputObjectMeta;
 
-use crate::ClusterError;
 use crate::helm::{HelmClient, Chart, InstalledChart};
 use crate::check::{
     CheckError, StatusCheck, InstallCheck, HelmVersion, AlreadyInstalled, SysChart, LoadableConfig,
     LoadBalancer,
 };
+use crate::error::K8InstallError;
+use crate::ClusterError;
 
 pub(crate) const DEFAULT_NAMESPACE: &str = "default";
 pub(crate) const DEFAULT_HELM_VERSION: &str = "3.3.4";
@@ -113,8 +114,8 @@ impl ClusterInstallerBuilder {
     pub fn build(self) -> Result<ClusterInstaller, ClusterError> {
         Ok(ClusterInstaller {
             config: self,
-            kube_client: K8Client::default()?,
-            helm_client: HelmClient::new()?,
+            kube_client: K8Client::default().map_err(K8InstallError::K8ClientError)?,
+            helm_client: HelmClient::new().map_err(K8InstallError::HelmError)?,
         })
     }
 
@@ -594,13 +595,15 @@ impl ClusterInstaller {
 
     /// Get all the available versions of fluvio chart
     pub fn versions() -> Result<Vec<Chart>, ClusterError> {
-        let helm_client = HelmClient::new()?;
-        let versions = helm_client.versions(DEFAULT_CHART_APP_NAME)?;
+        let helm_client = HelmClient::new().map_err(K8InstallError::HelmError)?;
+        let versions = helm_client
+            .versions(DEFAULT_CHART_APP_NAME)
+            .map_err(K8InstallError::HelmError)?;
         Ok(versions)
     }
 
     /// Get installed system chart
-    pub fn sys_charts() -> Result<Vec<InstalledChart>, ClusterError> {
+    pub fn sys_charts() -> Result<Vec<InstalledChart>, K8InstallError> {
         let helm_client = HelmClient::new()?;
         let sys_charts = helm_client.get_installed_chart_by_name(DEFAULT_CHART_SYS_REPO)?;
         Ok(sys_charts)
@@ -637,7 +640,9 @@ impl ClusterInstaller {
                         println!("✔️  {}", msg.green());
                     }
                     // unrecoverable error occured, return error
-                    StatusCheck::NotWorkingNoRemediation(failure) => return Err(failure.into()),
+                    StatusCheck::NotWorkingNoRemediation(failure) => {
+                        return Err(K8InstallError::PreCheck(failure).into());
+                    }
                     // recoverable error occured, try to fix
                     StatusCheck::NotWorking(failure, _) => {
                         match self.pre_install_fix(failure).await {
@@ -645,18 +650,18 @@ impl ClusterInstaller {
                                 println!("✔️  success");
                             }
                             // not able to recover, return error
-                            Err(err) => return Err(err),
+                            Err(err) => return Err(err.into()),
                         };
                     }
                 },
-                Err(err) => return Err(err.into()),
+                Err(err) => return Err(K8InstallError::PreCheck(err).into()),
             };
         }
 
         Ok(())
     }
 
-    async fn _try_minikube_tunnel(&self) -> Result<(), ClusterError> {
+    async fn _try_minikube_tunnel(&self) -> Result<(), K8InstallError> {
         let log_file = File::create("/tmp/tunnel.out")?;
         let error_file = log_file.try_clone()?;
 
@@ -672,7 +677,7 @@ impl ClusterInstaller {
 
     /// Given a pre-check error, attempt to automatically correct it
     #[instrument(skip(self, error))]
-    pub(crate) async fn pre_install_fix(&self, error: CheckError) -> Result<(), ClusterError> {
+    pub(crate) async fn pre_install_fix(&self, error: CheckError) -> Result<(), K8InstallError> {
         // Depending on what error occurred, try to fix the error.
         // If we handle the error successfully, return Ok(()) to indicate success
         // If we cannot handle this error, return it to bubble up
@@ -713,9 +718,9 @@ impl ClusterInstaller {
                     self.install_app()?;
                 }
                 // If Fluvio is already installed, skip install step
-                Err(ClusterError::PreCheckError {
-                    source: CheckError::AlreadyInstalled,
-                }) => {
+                Err(ClusterError::InstallK8(K8InstallError::PreCheck(
+                    CheckError::AlreadyInstalled,
+                ))) => {
                     debug!("Fluvio is already installed. Getting SC address");
                     let sc_address = self.wait_for_sc_service(&self.config.namespace).await?;
                     return Ok(sc_address);
@@ -729,19 +734,11 @@ impl ClusterInstaller {
         }
 
         let namespace = &self.config.namespace;
-        let sc_address = match self.wait_for_sc_service(namespace).await {
-            Ok(addr) => {
-                info!(addr = &*addr, "Fluvio SC is up");
-                addr
-            }
-            Err(err) => {
-                warn!("Unable to detect Fluvio service. If you're running on Minikube, make sure you have the tunnel up!");
-                return Err(ClusterError::Other(format!(
-                    "Unable to detect Fluvio service: {}",
-                    err
-                )));
-            }
-        };
+        let sc_address = self
+            .wait_for_sc_service(namespace)
+            .await
+            .map_err(|_| K8InstallError::UnableToDetectService)?;
+        info!(addr = %sc_address, "Fluvio SC is up:");
 
         if self.config.save_profile {
             self.update_profile(sc_address.clone())?;
@@ -771,7 +768,7 @@ impl ClusterInstaller {
     // TODO: Discussion at https://github.com/infinyon/fluvio/issues/235
     #[doc(hidden)]
     #[instrument(skip(self))]
-    pub fn _install_sys(&self) -> Result<(), ClusterError> {
+    pub fn _install_sys(&self) -> Result<(), K8InstallError> {
         let install_settings = &[("cloud", &*self.config.cloud)];
         match &self.config.chart_location {
             ChartLocation::Remote(chart_location) => {
@@ -813,7 +810,7 @@ impl ClusterInstaller {
 
     /// Install Fluvio Core chart on the configured cluster
     #[instrument(skip(self))]
-    fn install_app(&self) -> Result<(), ClusterError> {
+    fn install_app(&self) -> Result<(), K8InstallError> {
         debug!(
             "Installing fluvio with the following configuration: {:#?}",
             &self.config
@@ -862,9 +859,9 @@ impl ClusterInstaller {
                     .helm_client
                     .chart_version_exists(&self.config.chart_name, &self.config.chart_version)?
                 {
-                    return Err(ClusterError::Other(format!(
-                        "{}:{} not found in helm repo",
-                        &self.config.chart_name, &self.config.chart_version,
+                    return Err(K8InstallError::HelmChartNotFound(format!(
+                        "{}:{}",
+                        &self.config.chart_name, &self.config.chart_version
                     )));
                 }
                 debug!(
@@ -913,7 +910,7 @@ impl ClusterInstaller {
 
     /// Looks up the external address of a Fluvio SC instance in the given namespace
     #[instrument(skip(self, ns))]
-    async fn discover_sc_address(&self, ns: &str) -> Result<Option<String>, ClusterError> {
+    async fn discover_sc_address(&self, ns: &str) -> Result<Option<String>, K8InstallError> {
         use k8_client::http::status::StatusCode;
 
         let result = self
@@ -927,7 +924,7 @@ impl ClusterInstaller {
                 info!("no SC service found");
                 return Ok(None);
             }
-            Err(err) => return Err(ClusterError::from(err)),
+            Err(err) => return Err(K8InstallError::from(err)),
         };
 
         let ingress_addr = svc
@@ -952,7 +949,7 @@ impl ClusterInstaller {
 
     /// Wait until the Fluvio SC public service appears in Kubernetes
     #[instrument(skip(self, ns))]
-    async fn wait_for_sc_service(&self, ns: &str) -> Result<String, ClusterError> {
+    async fn wait_for_sc_service(&self, ns: &str) -> Result<String, K8InstallError> {
         info!("waiting for SC service");
         for i in 0..12 {
             if let Some(sock_addr) = self.discover_sc_address(ns).await? {
@@ -969,11 +966,11 @@ impl ClusterInstaller {
             sleep(Duration::from_millis(sleep_ms)).await;
         }
 
-        Err(ClusterError::SCServiceTimeout)
+        Err(K8InstallError::SCServiceTimeout)
     }
 
     /// Wait until the Fluvio SC public service appears in Kubernetes
-    async fn wait_for_sc_port_check(&self, sock_addr_str: &str) -> Result<(), ClusterError> {
+    async fn wait_for_sc_port_check(&self, sock_addr_str: &str) -> Result<(), K8InstallError> {
         info!(sock_addr = %sock_addr_str, "waiting for SC port check");
         for i in 0..12 {
             let sock_addr = self.wait_for_sc_dns(&sock_addr_str).await?;
@@ -986,14 +983,14 @@ impl ClusterInstaller {
             sleep(Duration::from_millis(sleep_ms)).await;
         }
         error!(sock_addr = %sock_addr_str, "timeout for SC port check");
-        Err(ClusterError::SCPortCheckTimeout)
+        Err(K8InstallError::SCPortCheckTimeout)
     }
 
     /// Wait until the Fluvio SC public service appears in Kubernetes
     async fn wait_for_sc_dns(
         &self,
         sock_addr_string: &str,
-    ) -> Result<Vec<SocketAddr>, ClusterError> {
+    ) -> Result<Vec<SocketAddr>, K8InstallError> {
         info!("waiting for SC dns resolution: {}", sock_addr_string);
         for i in 0..12 {
             match resolve(sock_addr_string).await {
@@ -1013,12 +1010,12 @@ impl ClusterInstaller {
         }
 
         error!("timedout sc dns: {}", sock_addr_string);
-        Err(ClusterError::SCDNSTimeout)
+        Err(K8InstallError::SCDNSTimeout)
     }
 
     /// Wait until all SPUs are ready and have ingress
     #[instrument(skip(self, ns))]
-    async fn wait_for_spu(&self, ns: &str, spu: u16) -> Result<bool, ClusterError> {
+    async fn wait_for_spu(&self, ns: &str, spu: u16) -> Result<bool, K8InstallError> {
         info!("waiting for SPU");
         for i in 0..12 {
             debug!("retrieving spu specs");
@@ -1053,7 +1050,7 @@ impl ClusterInstaller {
             }
         }
 
-        Err(ClusterError::SPUTimeout)
+        Err(K8InstallError::SPUTimeout)
     }
 
     /// Install server-side TLS by uploading secrets to kubernetes
@@ -1099,7 +1096,7 @@ impl ClusterInstaller {
     }
 
     /// Updates the Fluvio configuration with the newly installed cluster info.
-    fn update_profile(&self, external_addr: String) -> Result<(), ClusterError> {
+    fn update_profile(&self, external_addr: String) -> Result<(), K8InstallError> {
         debug!("updating profile for: {}", external_addr);
         let mut config_file = ConfigFile::load_default_or_new()?;
         let config = config_file.mut_config();
@@ -1134,12 +1131,12 @@ impl ClusterInstaller {
     }
 
     /// Determines a profile name from the name of the active Kubernetes context
-    fn compute_profile_name(&self) -> Result<String, ClusterError> {
+    fn compute_profile_name(&self) -> Result<String, K8InstallError> {
         let k8_config = K8Config::load()?;
 
         let kc_config = match k8_config {
             K8Config::Pod(_) => {
-                return Err(ClusterError::Other(
+                return Err(K8InstallError::Other(
                     "Pod config is not valid here".to_owned(),
                 ))
             }
@@ -1149,7 +1146,7 @@ impl ClusterInstaller {
         kc_config
             .config
             .current_context()
-            .ok_or_else(|| ClusterError::Other("No context fount".to_owned()))
+            .ok_or_else(|| K8InstallError::Other("No context fount".to_owned()))
             .map(|ctx| ctx.name.to_owned())
     }
 
@@ -1158,7 +1155,7 @@ impl ClusterInstaller {
         skip(self, cluster),
         fields(cluster_addr = &*cluster.addr)
     )]
-    async fn create_managed_spu_group(&self, cluster: &FluvioConfig) -> Result<(), ClusterError> {
+    async fn create_managed_spu_group(&self, cluster: &FluvioConfig) -> Result<(), K8InstallError> {
         debug!("trying to create managed spu: {:#?}", cluster);
         let name = self.config.group_name.clone();
         let fluvio = Fluvio::connect_with_config(cluster).await?;
