+++ src/extension-runner-local/src/error.rs
@@ -5,10 +5,7 @@ pub type Result<T> = std::result::Result<T, RunnerError>;
 #[derive(thiserror::Error, Debug)]
 pub enum RunnerError {
     #[error(transparent)]
-    IoError {
-        #[from]
-        source: IoError,
-    },
+    IoError(#[from] IoError),
     #[error("Unknown error: {0}")]
     Other(String),
 }
